requirements.txt
fastapi==0.95.2
uvicorn==0.24.0
pydantic==1.10.13
kafka-python==2.0.2
psycopg2-binary==2.9.10
redis==6.2.0
requests==2.32.4
python-multipart==0.0.20
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
prometheus-client==0.19.0
structlog==23.2.0 
firebase-admin==6.4.0 
google-cloud-firestore>=2.0.0
firebase-admin>=5.2.0
grpcio>=1.51.0
grpcio-tools>=1.51.0
exceptiongroup>=1.0.0 